JIT: move state body evaluation out-of-line
After:
```
Benchmark 1: ./mine-highlight.out benchmark/resources/llvm/llvm-8000.txt
  Time (mean ± σ):      21.0 ms ±   0.5 ms    [User: 20.1 ms, System: 0.8 ms]
  Range (min … max):    20.2 ms …  22.7 ms    142 runs

Benchmark 1: ./mine.out benchmark/resources/blocklist/blocklist-20000.txt
  Time (mean ± σ):      27.1 ms ±   0.6 ms    [User: 26.2 ms, System: 0.7 ms]
  Range (min … max):    26.2 ms …  29.9 ms    111 runs
```

+++ src/private/jit.hpp
@@ -69,8 +69,7 @@ public:
 
 struct RegexCompiledImpl {
   using EntryPointFn = bool (*)(Codepoint, evaluation::IndexType,
-                                evaluation::IndexType, void *, void *,
-                                unsigned);
+                                evaluation::IndexType, void *, void *, size_t);
   std::unique_ptr<RegexGraphImpl> m_graph;
   jit::ExecutableSection m_executable;
   EntryPointFn m_entrypoint;
